refactoring camera in external module [#76]
- external camera 기능의 모듈 내 위치 변경
