+++ external/src/main/java/com/foke/together/external/camera/internal/AndroidCameraDataSource.kt
@@ -1,4 +1,4 @@
-package com.foke.together.external.local
+package com.foke.together.external.camera.internal
 
 class AndroidCameraDataSource {
     // TODO: add android camera api here
